+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.28.0
+
+* Add --preserve-point-density-threshold option to reduce blank areas of the map at low zooms
+* Fix tile-join bug where use of --read-from would also accidentally enable --quiet
+
 # 2.27.0
 
 * Do more of line simplification in integer coordinates, to make behavior consistent across platforms